New auth controller and roles added to users
