+++ ActiLink/ActiLink/Organizers/Users/Service/UserService.cs
@@ -1,5 +1,6 @@
 ﻿using ActiLink.Configuration;
 using ActiLink.Organizers.Authentication;
+using ActiLink.Organizers.Authentication.Extensions;
 using ActiLink.Shared.Repositories;
 using ActiLink.Shared.ServiceUtils;
 using Microsoft.AspNetCore.Identity;
@@ -12,9 +13,7 @@ namespace ActiLink.Organizers.Users.Service
         private readonly IUnitOfWork _unitOfWork;
         private readonly UserManager<Organizer> _userManager;
         private static readonly string[] InvalidLoginError = ["Invalid email or password."];
-        private static readonly string[] InvalidRefreshTokenError = ["Invalid refresh token."];
         private static readonly string[] FailedRefreshTokenSave = ["Failed to save the refresh token."];
-        private static readonly string[] FailedRefreshTokenUpdate = ["Failed to update the refresh token."];
         private readonly JwtTokenProvider _tokenProvider;
         private readonly JwtSettings _jwtSettings;
         public UserService(IUnitOfWork unitOfWork, UserManager<Organizer> userManager, JwtTokenProvider provider, IOptions<JwtSettings> jwtOptions)
@@ -80,46 +79,6 @@ namespace ActiLink.Organizers.Users.Service
                 return GenericServiceResult<(string, string)>.Failure(FailedRefreshTokenSave);
 
             return GenericServiceResult<(string, string)>.Success((accessToken, refreshToken));
-
-        }
-
-        /// <summary>
-        /// Refreshes the access token using the specified <paramref name="refreshToken"/>.
-        /// </summary>
-        /// <param name="refreshToken"></param>
-        /// <returns>
-        /// The <see cref="Task"/> that represents the asynchronous operation, containing the <see cref="GenericServiceResult{T}"/> of the operation 
-        /// with the new access and refresh tokens or null and error messages if the refresh token is invalid or expired.
-        /// </returns>
-        public async Task<GenericServiceResult<(string AccessToken, string RefreshToken)>> RefreshTokenAsync(string refreshToken)
-        {
-            var refreshTokenEntity = await _unitOfWork
-                .RefreshTokenRepository
-                .GetValidTokenWithOwnerAsync(refreshToken);
-
-            if (refreshTokenEntity == null)
-                return GenericServiceResult<(string, string)>.Failure(InvalidRefreshTokenError);
-
-            var user = refreshTokenEntity.TokenOwner;
-
-            var newAccessToken = _tokenProvider.GenerateAccessToken(user);
-            var newRefreshToken = _tokenProvider.GenerateRefreshToken(user.Id!);
-
-            _unitOfWork.RefreshTokenRepository.Delete(refreshTokenEntity);
-
-            var newToken = new RefreshToken
-            {
-                Token = newRefreshToken,
-                ExpiryTimeUtc = DateTime.UtcNow.AddDays(_jwtSettings.RefreshTokenExpiryDays),
-                TokenOwner = user
-            };
-
-            await _unitOfWork.RefreshTokenRepository.AddAsync(newToken);
-            var saveResult = await _unitOfWork.SaveChangesAsync();
-            if (saveResult == 0)
-                return GenericServiceResult<(string, string)>.Failure(FailedRefreshTokenUpdate);
-
-            return GenericServiceResult<(string, string)>.Success((newAccessToken, newRefreshToken));
         }
 
         /// <summary>
